Additional checks added. Open map issue

+++ libft/ft_strlen.c
@@ -13,7 +13,7 @@
 //#include <stdio.h>
 #include "libft.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
 	int	i;
