+++ libft/ft_substr.c
@@ -13,7 +13,7 @@
 //#include <stdio.h>
 #include "libft.h"
 
-char	*ft_substr(char const *s, unsigned int start, size_t len)
+char	*ft_substr(char const *s, int start, int len)
 {
 	char	*substr;
